srcs/requirements/bonus/website/requirements.txt
Django==3.2
djangorestframework
requests